refactor: convert to library+binary project
create a lib.rs so that we can access modules from the unit
tests as well as the main.rs
